+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-813](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-813)
   [GEOD-849](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-849)
   [GEOD-867](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-867)
+- Filter parameter using the configuration file
+  [GEOD-881](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-881)
 - Add Content for the Exploratory Analysis Page
   [GLGVO-274](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-274)
-